equals() and hashCode() for generated EventHandlers
As Otto library uses EventHandler's equals() and hashCode() methods, they should be
implemented. For anonymous classes it's enough to check class names as each anonymous
class will get a unique name. However, subscribers of the same type might be
registered several times and additional "subscribers equal" check should be used.

As Otto Compiler is a compile-time dependency GeneratedEventHandler must be written to
the output during a round of annotation processing.